autodebug/requirements.txt
python-dotenv==1.0.0
requests==2.25.1
pyyaml==6.0